ci(gha): Update mihomo install method
This commit updates the method of installing mihomo in the GHA workflow.

The previous method used cURL to download a specific version of mihomo and then extract it. This method is now outdated and no longer works.

The new method uses wget to download the latest version of mihomo and then extract it. This method is more reliable and up-to-date.